+++ modules/plug_example/src/Plugin/name/John.php
@@ -13,7 +13,7 @@ namespace Drupal\plug_example\Plugin\name;
  *
  * @Name(
  *   id = "john",
- *   company = FALSE
+ *   company = TRUE
  * )
  */
 class John extends NameBase implements NameInterface {
